// src/domain/ledger/shared/types.ts
export enum AccountCode {
  // TODO: rename - we should save Settlement* for accounts that get used in settlement
  Settlement_Balance = 10100,
  Deposit = 10200,
  Unrestricted = 20100,
  Clearing_Credit = 20101,
  Restricted = 20200,
  Reserved = 20300,
  Committed_Outgoing = 20400,
  Dfsp = 60100,
  Net_Debit_Cap = 60200,
  Net_Debit_Cap_Control = 60201,
  Dev_Null = 60300,
  Clearing_Setup = 60400,
  Clearing_Limit = 60500,
  Unrestricted_Lock = 60600,
  Settlement_Outgoing = 60701,
  Settlement_Incoming = 60702,

  // TODO(LD): remove me! 
  TIMEOUT = 9000,
}

export enum TransferCode {
  Deposit = 10001,
  Withdraw = 20001,
  Clearing_Reserve = 30001,
  Clearing_Active_Check = 30002,
  Clearing_Fulfil = 30003,
  Clearing_Credit = 30004,
  Clearing_Reverse = 30005,
  Settlement_Deposit_Reduce = 40001,
  Settlement_Deposit_Increase = 40002,
  Net_Debit_Cap_Lock = 50001,
  Net_Debit_Cap_Sweep_To_Restricted = 50002,
  Net_Debit_Cap_Set_Limited = 50004,
  Net_Debit_Cap_Set_Unlimited = 50005,
  Net_Debit_Cap_Sweep_To_Unrestricted = 50006,
  Close_Account = 50007,
}